Expose colour conversion via Internal module
For background on where this came from, see https://github.com/georgefst/lifx-manager/issues/1.
